# src/main/java/it/bitify/libreria/model/entity/Card.java
package it.bitify.libreria.model.entity;

import jakarta.persistence.*;

import java.time.LocalDate;

@Entity
@Table(name = "Card")
public class Card {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @Column(name = "id")
    private Long id;

    @Column(name = "releaseDate", nullable = false)
    private LocalDate releaseDate;

    @OneToOne(mappedBy = "card")
    private Student student;

    public void setReleaseDate(LocalDate releaseDate) {
        this.releaseDate = releaseDate;
    }

    public void setIdCard(Long id) {
        this.id = id;
    }

    public LocalDate getReleaseDate() {
        return releaseDate;
    }

    public Long getIdCard() {
        return id;
    }
}