# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Libreria API definition
  version: v1
servers:
- url: http://localhost:8080
  description: base server url
...
paths:
  /api/v1/card:
    get:
      tags:
      - get-all
      - Card
      summary: Retrieve all cards
      description: Retrieve all card objects. The response is a list of card objects.
      operationId: getAllTessere
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Retrieved all cards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageCard"
    put:
      tags:
      - put
      - Card
      summary: Modify card
      description: Modify a card object by specifying its id
      operationId: editCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Card"
        required: true
      responses:
        "200":
          description: Card modified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          description: Id type mismatch
        "404":
          description: Card not found
        "500":
          description: Data integrity violation
    post:
      tags:
      - post
      - Card
      summary: Add card
      description: Add a card object to the database
      operationId: addCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Card"
        required: true
      responses:
        "200":
          description: Card added to the database
        "500":
          description: Data integrity violation
  /api/v1/card/{id}:
    get:
      tags:
        - get
        - Card
      summary: Retrieve card by id
      description: Retrieve a card object by specifying its id. The response is a
        card object
      operationId: getCardById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Found the card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          description: Id type mismatch
        "404":
          description: Card not found
    delete:
      tags:
        - delete
        - Card
      summary: Delete by id
      description: Delete a card object by specifying its id
      operationId: removeCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Card deleted
        "400":
          description: Id type mismatch
        "404":
          description: card not found
        "500":
          description: Data integrity violation
...
  /api/v1/book:
    get:
      tags:
      - Book
      - get-all
      summary: Retrieve all books
      description: Retrieve all Book objects. The response is a list of Book objects
      operationId: getAll_2
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Retrieved all Book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
    put:
      tags:
      - Book
      - put
      summary: Modify Book
      description: Modify a Book object by specifying its id
      operationId: 
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Book"
        required: true
      responses:
        "200":
          description: Book modified
        "400":
          description: Id type mismatch
        "404":
          description: Book not found
        "500":
          description: Data integrity violation
    post:
      tags:
      - post
      - Book
      summary: Add Book
      description: Add a Book object to the database
      operationId: addBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Book"
        required: true
      responses:
        "200":
          description: Book added to the database
        "500":
          description: Data integrity violation
  /api/v1/book/{id}:
    get:
      tags:
        - Book
        - get
      summary: Retrieve Book by id
      description: Retrieve a Book object by specifying its id. The response is a
        Book object
      operationId: getBookById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Book found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          description: Id type mismatch
        "404":
          description: Book not found
    delete:
      tags:
        - Book
        - delete
      summary: Delete Book by id
      description: Delete a Book object by specifying its id
      operationId: deleteBook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Book deleted
        "400":
          description: Id type mismatch
        "404":
          description: Book not found
        "500":
          description: Data integrity violation
  /api/v1/book/by-year-greater-than:
    get:
      tags:
        - get-all
        - Book
      summary: Retrieve books by specifying a publish year.
      description: Retrieve a list of book objects published after the specified year
      operationId: getByPublishYearGreaterThan
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Book found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageBook"
        "400":
          description: id type mismatch
  /api/v1/book/by-category-name:
    get:
      tags:
        - get-all
        - Book
      summary: Retrieve books by specifying a category.
      description: Retrieve a list of book objects belonging to a category.
      operationId: getByCategoryName
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Book found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageBook"
        "400":
          description: id type mismatch
  /api/v1/book/without-loan:
    get:
      tags:
        - get-all
        - Book
      summary: Retrieve non-borrowed books.
      description: Retrieve a list of non-borrowed book objects
      operationId: getWithoutLoan
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Book found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageBook"
  /api/v1/book/loan-interval:
    get:
      tags:
        - get-all
        - Book
      summary: Retrieve books by specifying a loan start date
      description: Retrieve a list of book objects lent after the specified date
      operationId: getByLoanInterval
      parameters:
        - name: start
          in: query
          required: true
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Book found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageBook"
        "400":
          description: id type mismatch
  /api/v1/book/loan-above-avg:
    get:
      tags:
        - get-all
        - Book
      summary: Retrieve books with above average number of loans.
      description: Retrieve a list of book objects with above average number of loans
      operationId: getByLoanAboveAvg
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/AscParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        "200":
          description: Book found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageBook"
...
components:
  schemas:
    Card:
      type: object
      properties:
        id:
          type: integer
          format: int64
        releaseDate:
          type: string
          format: date
...
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        books:
          type: array
          items:
            $ref: "#/components/schemas/Book"
    Book:
      type: object
      properties:
        id:
          type: integer
          format: int64
        year:
          type: integer
          format: int32
        category:
          $ref: "#/components/schemas/Category"
        title:
          type: string
        author:
          type: string
...
    PageBook:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Book'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
    PageCard:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Card'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
...
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 5
    AscParam:
      name: ascending
      in: query
      required: false
      schema:
        type: boolean
        default: true
    SortByParam:
      name: sortBy
      in: query
      required: false
      schema:
        type: string
        default: id